Refactor: Change export type
